+++ ipyslides/shared_vars.py
@@ -1,4 +1,4 @@
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 
 _md_extensions = ['fenced_code','tables','codehilite','footnotes'] # For MArkdown Parser